TinyTestServer/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
SQLAlchemy
psycopg2
ngrok-api
